requirements.txt
asgiref==3.9.1
boto3==1.40.0
botocore==1.40.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-appconf==1.1.0
django-imagekit==5.0.0
django-storages==1.14.6
djangorestframework==3.16.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pilkit==3.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
